move test to standalone file; inspect.test.ts has pre-existing localhost failures on main
